# Conflicts:
#	drivers/can/can_mcan.c
#	drivers/dma/dma_stm32.c
#	drivers/serial/uart_stm32.c
#	drivers/serial/uart_stm32.h
#	include/zephyr/drivers/can/can_mcan.h
#	include/zephyr/kernel/thread.h
#	kernel/poll.c
#	kernel/sched.c

+++ drivers/clock_control/clock_stm32_ll_h7.c
@@ -222,7 +222,7 @@ static uint32_t get_pllout_frequency(uint32_t pllsrc_freq,
 
     __ASSERT_NO_MSG((pllm_div != 0) && (pllout_div != 0));
 
-    return PLLX_OUT_FREQ(pllsrc_freq, pllm_div, plln_mul_u64, plln_frac, pllout_div);
+    return (uint32_t)PLLX_OUT_FREQ(pllsrc_freq, pllm_div, plln_mul_u64, plln_frac, pllout_div);
 }
 
 __unused
